backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# ML/NLP Dependencies
torch>=2.0.0
transformers>=4.35.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
sentencepiece>=0.1.99

# Medical NLP
spacy==3.7.2
scispacy==0.5.3
# Run after install: python -m spacy download en_core_sci_md

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# OCR and Document Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
easyocr==1.7.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
python-docx==1.1.0

# AI API
groq>=0.4.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
